2017/server/requirements.txt
backports-abc==0.5
certifi==2017.11.5
pkg-resources==0.0.0
pyserial==3.4
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.2
